samples/ticket-classification/pyproject.toml
[project]
name = "ticket-classification"
version = "0.0.5"
description = "Ticket classification system"
authors = [{ name = "Cristi Pufu", email = "cristian.pufu@uipath.com" }]
dependencies = [
    "langgraph>=0.2.70",
    "langchain-core>=0.3.34",
    "langchain-anthropic>=0.3.7",
    "langgraph-checkpoint-sqlite>=2.0.3",
    "python-dotenv>=1.0.1",
    "uipath_sdk==0.0.27",
    "pydantic>=2.10.6",
    "aiohttp>=3.11.12",
    "typing-extensions>=4.12.2",
    "ipython>=8.32.0"
]
requires-python = ">=3.10"
